Part 1 - Data Collection
Data collection on SpaceX Falcon 9 first stage Landing Prediction. We collect information on the outcome of the landing, the type of the landing, number of flights with that core, whether gridfins were used, wheter the core is reused, wheter legs were used, the landing pad used, the block of the core which is a number used to seperate version of cores, the number of times this specific core has been reused, and the serial of the core. Then we prepare it for further analysis.